+++ crates/server/src/http/mod.rs
@@ -1,0 +1,2 @@
+pub mod range;
+pub mod responses;
