+++ crates/server/src/services/mod.rs
@@ -1,0 +1,4 @@
+pub mod artwork;
+pub mod hls_cache;
+pub mod playback;
+pub mod track_duration;
